Working over CCITTFaxCodec: attempt to use jai-imageio code

+++ src/main/java/net/algart/matrices/tiff/TiffIFD.java
@@ -389,7 +389,7 @@ public class TiffIFD {
         return getValue(tag, Number.class).orElse(defaultValue).longValue();
     }
 
-    public long optLong(int tag, int defaultValue) {
+    public long optLong(int tag, long defaultValue) {
         return optValue(tag, Number.class).orElse(defaultValue).longValue();
     }
